# swagger.yaml
openapi: 3.0.0
info:
  title: Pantreazy
  description: Food/Recipe App
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
  /accounts/authenticate:
    post:
      summary: Authenticate account credentials and return a JWT token and a cookie with a refresh token
      description: Accounts must be verified before authenticating.
      operationId: authenticate
      tags: 
        - accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "jason@example.com"
                password:
                  type: string
                  example: "pass123"
              required:
                - email
                - password
      responses:
        "200":
          description: Account details, a JWT access token and a refresh token cookie
          headers:
            Set-Cookie:
              description: "`refreshToken`"
              schema:
                type: string
                example: refreshToken=51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550; Path=/; Expires=Tue, 16 Jun 2020 09:14:17 GMT; HttpOnly
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  title:
                    type: string
                    example: "Mr"
                  firstName:
                    type: string
                    example: "Jason"
                  lastName:
                    type: string
                    example: "Watmore"
                  email:
                    type: string
                    example: "jason@example.com"
                  role:
                    type: string
                    example: "Admin"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  isVerified:
                    type: boolean
                    example: true
                  jwtToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y"
        "400":
          description: The email or password is incorrect
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email or password is incorrect"
...
  /foods/{barcode}:
    parameters:
        - in: path
          name: barcode
          description: food barcode. 
          required: true
          example: "072250020756"
          schema:
            type: string
    get:
      summary: get a specific food item
      description: 
      operationId: 
      tags: 
        - foods
      responses:
        "200":
          description: An object with food info
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    barcode:
                      type: string
                      example: "072250020756"
                    product_name:
                      type: string
                      example: "Thick sliced multigrain bread"
                    food_group:
                      type: string
                      example: "Jason"
                    food_group_tags:
                      type: array
                      items:
                        type: string
                        example: ["en:cereals-and-potatoe", "en:bread"]
                    image:
                      type: string
                      example: "https://images.openfoodfacts.org/images/products/007/225/002/0756/front_en.8.200.jpg"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/InvalidInput"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"
    InvalidInput:
      description: Invalid Input
      content:
        application/json:
          schema:
            type: object 
            properties:
              message:
                type: string
                example: "Invalid Input"